Now working
- added fetch for start of historic values
- debug level back to info
- switched everything to double

+++ src/main/java/ch/furchert/iotapp/controller/DataController.java
@@ -50,11 +50,11 @@ public class DataController {
     }
 
     @GetMapping("/deviceStatus")
-    public ResponseEntity<int[]> getStatus(@RequestParam(value = "device", defaultValue = "terra1") String device) {
+    public ResponseEntity<double[]> getStatus(@RequestParam(value = "device", defaultValue = "terra1") String device) {
 
         log.debug("Getting statusdata from InfluxDB for device: {}", device);
 
-        int[] array = influxService.queryStatus(device);
+        double[] array = influxService.queryStatus(device);
 
         log.debug("die stati: {}", array);
         return new ResponseEntity<>(array, HttpStatus.OK);
